Add ChaCha20-Poly1305 AEAD encryption for botnet linking
Implements ChaCha20-Poly1305 authenticated encryption as the preferred
cipher for encrypted botnet links, replacing the legacy AES-ECB
(ghost+case3) method.

Key changes:
- New crypto/chacha_util.cc/h: OpenSSL EVP-based ChaCha20-Poly1305
  encrypt/decrypt with key derivation and base64 wire format
- New LINK_CHACHA20_POLY1305 enclink method with HMAC-SHA256 key
  rotation (chacha_rotate) for forward secrecy
- Downgrade-protected v2 link handshake: cipher list included in
  SHA1 hash to prevent MITM cipher downgrade attacks
- Per-link encryption type display in bot tree (enc_type field)
- Runtime cipher availability check (fatal if OpenSSL < 1.1.0)
- Legacy AES-ECB links log a hint to upgrade for ChaCha20-Poly1305
- Greeting-based encryption negotiation (moved from cont_link to
  dcc_bot_new for proper handling)

+++ src/dccutil.cc
@@ -98,8 +98,8 @@ colorbuf(char *buf, size_t len, int idx, size_t bufsiz)
   int cidx = coloridx(idx);
   int schar = 0;
   char buf3[1024] = "", buf2[15] = "", c = 0;
-  static int8_t stdout_cflags = 0;
-  int8_t *cflags;
+  static int stdout_cflags = 0;
+  int *cflags;
 
   if (idx == -1) {
     cflags = &stdout_cflags;
